+++ lib/kdialogs/src/show_bottom_alert.dart
@@ -8,6 +8,7 @@ Future<bool> showBottomAlertKDialog(
   bool retryable = false,
   String? acceptText,
   String? retryText,
+  bool useRootNavigator = true,
 }) async {
   title ??= strings.bottomErrorAlertTitle;
   acceptText ??= strings.acceptButtonText;
@@ -18,7 +19,7 @@ Future<bool> showBottomAlertKDialog(
     backgroundColor: Colors.transparent,
     useSafeArea: true,
     enableDrag: true,
-    useRootNavigator: true,
+    useRootNavigator: useRootNavigator,
     builder: (context) {
       return SafeArea(
         child: PopScope(
@@ -64,9 +65,9 @@ Future<bool> showBottomAlertKDialog(
                           Visibility(
                             visible: retryable,
                             child: TextButton(
-                              onPressed: () =>
-                                  Navigator.of(context, rootNavigator: true)
-                                      .pop(retryable),
+                              onPressed: () => Navigator.of(context,
+                                      rootNavigator: useRootNavigator)
+                                  .pop(retryable),
                               child: Text(retryText ?? "RETRY"),
                             ),
                           ),
@@ -74,7 +75,7 @@ Future<bool> showBottomAlertKDialog(
                           ElevatedButton(
                             onPressed: () => Navigator.of(
                               context,
-                              rootNavigator: true,
+                              rootNavigator: useRootNavigator,
                             ).pop(false),
                             child: Text(acceptText ?? "ACCEPT"),
                           )