+++ lib/kdialogs/src/show_kdialog_content.dart
@@ -22,6 +22,7 @@ Future<T?> showKDialogContent<T>(
   bool allowBackButtonToClose = true,
   Color? backgroundColor,
   required Widget Function(BuildContext context) builder,
+  bool useRootNavigator = false,
 }) async {
   title ??= strings.defaultDialogTitle;
   saveBtnText ??= strings.saveButtonText;
@@ -37,7 +38,10 @@ Future<T?> showKDialogContent<T>(
             children: [
               IconButton(
                 iconSize: 25,
-                onPressed: () => Navigator.of(context).pop(null),
+                onPressed: () => Navigator.of(
+                  context,
+                  rootNavigator: useRootNavigator,
+                ).pop(null),
                 icon: const Icon(Icons.close),
               ),
               Flexible(
@@ -63,8 +67,10 @@ Future<T?> showKDialogContent<T>(
               bool shouldClose =
                   await onSave(); // ✅ Esperamos el resultado de onSave
               if (shouldClose && context.mounted) {
-                Navigator.of(context)
-                    .pop(true); // ✅ Cierra el diálogo si onSave devuelve true
+                Navigator.of(
+                  context,
+                  rootNavigator: useRootNavigator,
+                ).pop(true); // ✅ Cierra el diálogo si onSave devuelve true
               }
             },
             style: Theme.of(context).textButtonTheme.style,
@@ -78,6 +84,7 @@ Future<T?> showKDialogContent<T>(
     context: context,
     barrierDismissible: closeOnOutsideTap,
     useSafeArea: true,
+    useRootNavigator: useRootNavigator,
     builder: (context) {
       return PopScope(
         canPop: allowBackButtonToClose,